Do not use `-r` option in gapd.sh
Using it resulted in the server running the version of the package
from the GAP latest release, not the one being tested.

It is possible to pass `-r` to `gapd.sh` via `$GAPOPTS` variable,
and in the future we may run tests with and without it to check
backwards compatibility.